feat(link): add /link to create an embed from a url
